# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Java Speaker Club API
  description: API для управления докладами, лекциями, ролями и медиафайлами спикеров
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/roles:
    get:
      tags: [Roles]
      summary: Список всех ролей
      operationId: findAllRoles
      responses:
        '200':
          description: Список ролей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'

    post:
      tags: [Roles]
      summary: Создать роль
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '201':
          description: Роль создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Спикер не найден или роль уже назначена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/roles/by-speaker/{speakerId}:
    get:
      tags: [Roles]
      summary: Роли спикера по его ID
      operationId: findRolesBySpeakerId
      parameters:
        - name: speakerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список ролей спикера
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
  /api/roles/{id}:
    delete:
      tags: [Roles]
      summary: Удалить роль
      operationId: deleteRole
      parameters:
        - $ref: '#/components/parameters/RoleId'
      responses:
        '204':
          description: Роль удалена
        '404':
          description: Роль не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ── Speakers ────────────────────────────────────────────────────────
  /api/speakers:
    get:
      tags: [Speakers]
      summary: Поиск спикеров по роли и NPS
      operationId: findSpeakersByRole
      parameters:
        - name: role
          in: query
          required: false
          description: Фильтр по названию роли
          schema:
            type: string
        - name: npsMin
          in: query
          required: false
          description: Минимальное значение NPS
          schema:
            type: number
            format: decimal
        - name: npsMax
          in: query
          required: false
          description: Максимальное значение NPS
          schema:
            type: number
            format: decimal
      responses:
        '200':
          description: Список спикеров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpeakerResponse'
  /api/speakers/{id}:
    get:
      tags: [Speakers]
      summary: Получить спикера по ID
      operationId: findSpeakerById
      parameters:
        - $ref: '#/components/parameters/SpeakerId'
      responses:
        '200':
          description: Спикер найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakerResponse'
        '404':
          description: Спикер не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/speakers/{id}/nps:
    put:
      tags: [Speakers]
      summary: Обновить NPS спикера
      operationId: updateNps
      parameters:
        - $ref: '#/components/parameters/SpeakerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NpsUpdateRequest'
      responses:
        '200':
          description: NPS обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakerResponse'
        '404':
          description: Спикер не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ── Media ───────────────────────────────────────────────────────────
...
components:
  parameters:
...
    RoleId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    SpeakerId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoleCreateRequest:
      type: object
      required: [name, speakerId]
      properties:
        name:
          type: string
          example: SPEAKER
        speakerId:
          type: integer
          format: int64
          example: 1
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: SPEAKER
        speakerId:
          type: integer
          format: int64
        createdDate:
          type: string
          format: date-time

    # ── Speakers ────────────────────────────────────────────────────
    SpeakerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          example: johndoe
        nps:
          type: number
          format: decimal
          nullable: true
          example: 8.5
        createdDate:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponse'
...
    NpsUpdateRequest:
      type: object
      required: [nps]
      properties:
        nps:
          type: number
          format: decimal
          example: 9.2
...
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: "Talk not found: 550e8400-e29b-41d4-a716-446655440000"
        instance:
          type: string
          format: uri